chore: implement RPC multicalls within token validation script
