+++ tokens/main.json
@@ -13,14 +13,6 @@
   "tokens": [
     {
       "chainId": 8453,
-      "address": "0xd52333441c0553fACb259600FA833a69186893A5",
-      "name": "PAAL AI",
-      "symbol": "$PAAL",
-      "decimals": 18,
-      "logoURI": "https://raw.githubusercontent.com/hydrexfi/hydrex-lists/main/assets/tokens/$PAAL.png"
-    },
-    {
-      "chainId": 8453,
       "address": "0xc5fecC3a29Fb57B5024eEc8a2239d4621e111CBE",
       "name": "1INCH Token",
       "symbol": "1INCH",
@@ -1280,6 +1272,14 @@
       "symbol": "OX",
       "decimals": 18,
       "logoURI": "https://raw.githubusercontent.com/hydrexfi/hydrex-lists/main/assets/tokens/OX.png"
+    },
+    {
+      "chainId": 8453,
+      "address": "0xd52333441c0553fACb259600FA833a69186893A5",
+      "name": "PAAL AI",
+      "symbol": "PAAL",
+      "decimals": 18,
+      "logoURI": "https://raw.githubusercontent.com/hydrexfi/hydrex-lists/main/assets/tokens/PAAL.png"
     },
     {
       "chainId": 8453,
